requirements.txt
openai-whisper>=20230314
paddlepaddle-gpu==2.6.2
paddleocr>=2.7.0
pygame>=2.0.0
opencv-python>=4.5.0
vosk>=0.3.45
pyaudio>=0.2.11
fastapi>=0.100.0
uvicorn>=0.23.0
websocket-client>=1.6.0
